[BUGFIX] Language comparison fix
TYPO3 13 changed behaviour in `DrawingConfiguration` in favour of a directly callable
`langaugeTranslationMode` instead of the dropped `languageMode`. On updating this
extension, this was not checked and therefore not handled.

With this commit the condition inside the `PageLayout` is adjusted respecting both
TYPO3 12 and 13 in handling the current DrawingConfiguration and rendering the
respective necessary Partial.